v2.1.0

+++ src/components/SettingsFolderButton.js
@@ -1,13 +1,20 @@
 import {React, useState} from 'react'
 import {Dropdown, DropdownButton, Modal, Button, Form, ListGroup} from "react-bootstrap"
 import {database, db} from "../firebase"
-import { deleteDoc, doc, updateDoc, arrayUnion, arrayRemove, collection, query, where, getDocs } from "firebase/firestore"
+import { deleteDoc, doc, updateDoc, arrayUnion, arrayRemove, collection, query, where, getDocs, getDoc } from "firebase/firestore"
+import rgbHex from 'rgb-hex';
 
 export default function SettingsFolderButton({folder}) {
     const [open, setOpen] = useState(false)
     const [settingsOpen, setSettingsOpen] = useState(false)
     const [email, setEmail] = useState("")
     const [name, setName] = useState("")
+    const [color, setColor] = useState("")
+
+
+    // const colorPick = document.getElementById("color-pick")
+    const folderColor = folder.color
+    // console.log(color)
 
     function openModal() {
         setOpen(true);
@@ -28,26 +35,39 @@ export default function SettingsFolderButton({folder}) {
     function handleDelete(e) {
         e.preventDefault();
 
+        if(folder.creator)
         deleteDoc(doc(db, "folders", folder.id));
         
         closeModal()
     }
 
-    function handleAddUser(e){
+    function handleSave(e){
         e.preventDefault()
         const folderRef = doc(db, "folders", folder.id);
         const fileRef = collection(db, "files");
 
         if(email != ""){
             updateDoc(folderRef, {
-                members: arrayUnion(email)
+                members: arrayUnion(email),
             })
-
             // const f = query(fileRef, where("folderId", "==", folder.id));
             
         }else{
-            alert("email must exist!")
+            
         }
+
+        if(name != ""){
+            updateDoc(folderRef, {
+                name: name
+            })
+        }
+        
+        if(color){
+            updateDoc(folderRef, {
+                color: color
+            })
+        }
+        
 
         closeSettingsModal()
     }
@@ -89,23 +109,30 @@ export default function SettingsFolderButton({folder}) {
             <Form>
             <Modal.Body>
                 <h4>{folder.name}</h4>
-                    <Form.Label>Folder name:</Form.Label>
-                    <Form.Control value={folder.name} onChange={e => setName(e.target.value)}></Form.Control>
-                    <Form.Label>Invite members to view this folder:</Form.Label>
-                    <Form.Control type="email" placeholder="type a valid email address..." onChange={e => setEmail(e.target.value)}></Form.Control>
+                    <div>
+                        <Form.Label>Folder name:</Form.Label>
+                        <Form.Control defaultValue={folder.name} onChange={e => setName(e.target.value)}></Form.Control>
+                        <Form.Label>Invite members to view this folder:</Form.Label>
+                        <Form.Control type="email" placeholder="type a valid email address..." onChange={e => setEmail(e.target.value)}></Form.Control>
+                    </div>
 
-                    <span>Members added:</span>
-                    <ListGroup className="members-list">
-                    {folder.members.map((child) => (
-                            <ListGroup.Item>{child}</ListGroup.Item>
-                    ))}
-                    </ListGroup>
+                    <div className="cp">
+                       <span>Color: </span> <input type="color" id="color-pick" defaultValue={folderColor} className="color-pick" onChange={e => setColor(e.target.value)}/>
+                    </div>
 
+                    <div>
+                        <span>Members added:</span>
+                        <ListGroup className="members-list">
+                        {folder.members.map((child) => (
+                                <ListGroup.Item>{child}</ListGroup.Item>
+                        ))}
+                        </ListGroup>
+                    </div>
                     
                 </Modal.Body>
                 <Modal.Footer>
                 <Button onClick={closeSettingsModal}>Close</Button>
-                <Button onClick={handleAddUser} variant='success' type="submit">Save</Button>
+                <Button onClick={handleSave} variant='success' type="submit">Save</Button>
             </Modal.Footer>
             </Form>
         </Modal>
